+++ CopilotKit/examples/next-openai/src/app/components/destination-table.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import React from "react";
 import { DestinationRow } from "./destination-row";
 import { Destination } from "./vacation-list";