+++ CopilotKit/packages/react-core/CHANGELOG.md
@@ -1,5 +1,107 @@
 # ui
 
+## 0.15.0
+
+### Minor Changes
+
+- 8a5cecd: only forward functions if non-empty
+- 87f1fa0: rebase
+- 15d4afc: debugging
+- c40a0d1: Filter out empty function descriptions
+- prep for chat protocol v2
+- bbd152e: backend sdks prep
+- 8517bb1: trying again
+- 478840a: carry function propagation fix to chat v2
+
+### Patch Changes
+
+- Updated dependencies [8a5cecd]
+- Updated dependencies [87f1fa0]
+- Updated dependencies [15d4afc]
+- Updated dependencies [c40a0d1]
+- Updated dependencies
+- Updated dependencies [bbd152e]
+- Updated dependencies [8517bb1]
+- Updated dependencies [478840a]
+  - @copilotkit/shared@0.1.0
+
+## 0.15.0-alpha.6
+
+### Minor Changes
+
+- rebase
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/shared@0.1.0-alpha.6
+
+## 0.15.0-alpha.5
+
+### Minor Changes
+
+- carry function propagation fix to chat v2
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/shared@0.1.0-alpha.5
+
+## 0.15.0-alpha.4
+
+### Minor Changes
+
+- only forward functions if non-empty
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/shared@0.1.0-alpha.4
+
+## 0.15.0-alpha.3
+
+### Minor Changes
+
+- debugging
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/shared@0.1.0-alpha.3
+
+## 0.15.0-alpha.2
+
+### Minor Changes
+
+- trying again
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/shared@0.1.0-alpha.2
+
+## 0.15.0-alpha.1
+
+### Minor Changes
+
+- Filter out empty function descriptions
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/shared@0.1.0-alpha.1
+
+## 0.15.0-alpha.0
+
+### Minor Changes
+
+- backend sdks prep
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/shared@0.1.0-alpha.0
+
 ## 0.14.0
 
 ### Minor Changes