+++ CopilotKit/packages/tsconfig/CHANGELOG.md
@@ -1,5 +1,60 @@
 # tsconfig
 
+## 0.6.0
+
+### Minor Changes
+
+- 8a5cecd: only forward functions if non-empty
+- 87f1fa0: rebase
+- 15d4afc: debugging
+- c40a0d1: Filter out empty function descriptions
+- prep for chat protocol v2
+- bbd152e: backend sdks prep
+- 8517bb1: trying again
+- 478840a: carry function propagation fix to chat v2
+
+## 0.6.0-alpha.6
+
+### Minor Changes
+
+- rebase
+
+## 0.6.0-alpha.5
+
+### Minor Changes
+
+- carry function propagation fix to chat v2
+
+## 0.6.0-alpha.4
+
+### Minor Changes
+
+- only forward functions if non-empty
+
+## 0.6.0-alpha.3
+
+### Minor Changes
+
+- debugging
+
+## 0.6.0-alpha.2
+
+### Minor Changes
+
+- trying again
+
+## 0.6.0-alpha.1
+
+### Minor Changes
+
+- Filter out empty function descriptions
+
+## 0.6.0-alpha.0
+
+### Minor Changes
+
+- backend sdks prep
+
 ## 0.5.0
 
 ### Minor Changes